keycloak introduced

+++ src/main/java/com/example/twitter/controller/TweetController.java
@@ -1,19 +1,18 @@
 package com.example.twitter.controller;
 
-import com.example.twitter.context.UserContextHolder;
 import com.example.twitter.controller.dto.CommentDto;
 import com.example.twitter.controller.dto.TweetCreateDto;
 import com.example.twitter.controller.dto.TweetGetDto;
 import com.example.twitter.controller.dto.TweetUpdateDto;
+import com.example.twitter.exception.NotValidOwnerException;
 import com.example.twitter.model.Comment;
 import com.example.twitter.model.Tweet;
-import com.example.twitter.model.User;
 import com.example.twitter.service.CommentService;
+import com.example.twitter.service.KeycloakService;
 import com.example.twitter.service.TweetService;
 import com.example.twitter.service.UserService;
 import com.example.twitter.utils.ObjectMapperUtils;
 import org.springframework.http.HttpStatus;
-import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
@@ -24,19 +23,32 @@ public class TweetController {
 
     private final TweetService tweetService;
     private final UserService userService;
+    private final KeycloakService keycloakService;
     private final CommentService commentService;
 
-    public TweetController(TweetService tweetService, UserService userService, CommentService commentService) {
+    public TweetController(TweetService tweetService, UserService userService, KeycloakService keycloakService, CommentService commentService) {
         this.tweetService = tweetService;
         this.userService = userService;
+        this.keycloakService = keycloakService;
         this.commentService = commentService;
     }
 
 
     @GetMapping("/v1/tweets")
     @ResponseStatus(HttpStatus.OK)
-    public List<TweetGetDto> getTweets() {
-        List<Tweet> tweets = tweetService.getTweets();
+    public List<TweetGetDto> getCurrentUserTweets() {
+
+        String userIdByToken = keycloakService.extractUserId();
+        List<Tweet> tweets = tweetService.getUserTweets(userIdByToken);
+        List<TweetGetDto> tweetsGetDto = ObjectMapperUtils.mapAll(tweets, TweetGetDto.class);
+        return tweetsGetDto;
+
+    }
+
+    @GetMapping("/v1/tweets/user/{id}")
+    @ResponseStatus(HttpStatus.OK)
+    public List<TweetGetDto> getUserTweets(@PathVariable("id") String userId) {
+        List<Tweet> tweets = tweetService.getUserTweets(userId);
         List<TweetGetDto> tweetsGetDto = ObjectMapperUtils.mapAll(tweets, TweetGetDto.class);
         return tweetsGetDto;
     }
@@ -49,51 +61,59 @@ public class TweetController {
     }
 
 
-    @PostMapping("v1/tweets")
+    @PostMapping("/v1/tweets")
     @ResponseStatus(HttpStatus.OK)
     public TweetGetDto addTweet(@Valid @RequestBody TweetCreateDto tweetCreateDto) {
         Tweet tweet = ObjectMapperUtils.map(tweetCreateDto, Tweet.class);
-        User user = UserContextHolder.get().getUser();
-        tweet.setUser(user);
+        String userIdByToken = keycloakService.extractUserId();
 
+        tweet.setUserId(userIdByToken);
         Tweet returnedTweet = tweetService.save(tweet);
 
         return ObjectMapperUtils.map(returnedTweet, TweetGetDto.class);
     }
 
-    @PutMapping("v1/tweets")
+    @PutMapping("/v1/tweets")
     @ResponseStatus(HttpStatus.OK)
     public TweetGetDto updateTweet(@Valid @RequestBody TweetUpdateDto tweetUpdateDto) {
         Tweet tweet = ObjectMapperUtils.map(tweetUpdateDto, Tweet.class);
-        User user = UserContextHolder.get().getUser();
-        if(user.getId() != tweet.getUser().getId()){
-            //TODO
-            // throw exception
+        Tweet existingTweet = tweetService.getTweetById(tweet.getId());
+
+        String userIdByToken = keycloakService.extractUserId();
+        if (!userIdByToken.equals(existingTweet.getUserId())) {
+            throw new NotValidOwnerException(existingTweet.getUserId(), "#NotValidOwner");
         }
-        Tweet returnedTweet = tweetService.update(tweet);
+
+        existingTweet.setText(tweet.getText());
+        Tweet returnedTweet = tweetService.update(existingTweet);
         return ObjectMapperUtils.map(returnedTweet, TweetGetDto.class);
+
     }
 
     @DeleteMapping("v1/tweet/{id}")
     @ResponseStatus(HttpStatus.OK)
     public void deleteTweet(@PathVariable String id) {
-        Tweet tweet = tweetService.getTweetById(Long.valueOf(id));
-        User user = UserContextHolder.get().getUser();
-        if(user.getId() != tweet.getUser().getId()){
-            //TODO
-            // throw exception
+
+        Tweet existingTweet = tweetService.getTweetById(Long.valueOf(id));
+
+        String userIdByToken = keycloakService.extractUserId();
+        if (!userIdByToken.equals(existingTweet.getUserId())) {
+            throw new NotValidOwnerException(existingTweet.getUserId(), "#NotValidOwner");
         }
-        tweetService.delete(tweet);
+
+        tweetService.delete(existingTweet);
     }
 
     @PostMapping("v1/tweet/comments")
     @ResponseStatus(HttpStatus.OK)
     public TweetGetDto addComment(@Valid @RequestBody CommentDto commentDto) {
+
+        String userIdByToken = keycloakService.extractUserId();
+
         Comment comment = ObjectMapperUtils.map(commentDto, Comment.class);
+        comment.setUserId(userIdByToken);
 
         Tweet tweet = tweetService.getTweetById(commentDto.getTweetId());
-        User user = UserContextHolder.get().getUser();
-        comment.setUser(user);
 
         tweetService.addComment(tweet, comment);
 
@@ -105,13 +125,16 @@ public class TweetController {
     @ResponseStatus(HttpStatus.OK)
     public void deleteComment(@PathVariable String id) {
 
+        String userIdByToken = keycloakService.extractUserId();
         Comment comment = commentService.getCommentById(Long.valueOf(id));
-        User user = UserContextHolder.get().getUser();
-        if(user.getId() != comment.getUser().getId()){
-            //TODO
-            // throw exception
+
+        if (!userIdByToken.equals(comment.getUserId())) {
+            throw new NotValidOwnerException(comment.getUserId(), "#NotValidOwner");
         }
+
         commentService.delete(comment);
 
     }
+
+
 }
